Add durable Session ownership and resume
Sessions previously had no stable identity, exclusive owner, or reconstructable conversation leaf. That left the owner loop unable to fence stale effects across process restarts.\n\nAdd a locked, checksummed Session store with bounded identity allocation, durable ownership epochs, and parent-linked conversation records. Fence owner drives and operation-journal effects through the current Session while preserving late completion recovery and the existing fixed byte budgets.\n\nRefs #5

+++ src/checkpoint_store.zig
@@ -22,7 +22,7 @@ pub fn publish(
     temp_path: []const u8,
     encoded: []u8,
     agent_id: u64,
-    generation: u64,
+    generation: u32,
     page: []const u8,
     fault: ?FaultHook,
 ) !void {
